splitted the "message store" which is for the smtp server and "readable message store" which should be used by the webapp for data retrieval, also made dbContext Transient globally
